refactor to compute embedded fields form script url
